### 2024-02-22 #6395 @editorialbot
[PRE REVIEW]: pyradiosky: A Python package for Radio Sky Models
**Submitting author:** bhazelton (<a href="http://orcid.org/0000-0001-7532-645X">Bryna J. Hazelton</a>)
**Repository:** https://github.com/RadioAstronomySoftwareGroup/pyradiosky
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.3.1

Thanks for submitting your paper to JOSS bhazelton. **Currently, there isn't a JOSS editor assigned** to your paper.

bhazelton if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
