### 2024-02-27 fc4dd1d
Some minor readme fixes
