+++ .github/workflows/externaltests.yaml
@@ -14,7 +14,7 @@ jobs:
         # Adding -l {0} helps ensure conda can be found properly.
         shell: bash -l {0}
     env:
-      ENV_NAME: pyuvsim_tests
+      ENV_NAME: pyuvsim_tests_openmpi
       PYTHON: "3.11"
 
     steps:
