### 2024-03-13 803f1a8
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/PyCQA/bandit: 1.7.7 → 1.7.8](https://github.com/PyCQA/bandit/compare/1.7.7...1.7.8)
