+++ .pre-commit-config.yaml
@@ -34,7 +34,7 @@ repos:
         - id: black
 
   - repo: https://github.com/PyCQA/bandit
-    rev: 1.7.7
+    rev: 1.7.8
     hooks:
       - id: bandit
         args: [--skip, "B101", --recursive, pyuvdata]
