### 2024-03-18 #6503 @editorialbot
[REVIEW]: pyradiosky: A Python package for Radio Sky Models
**Submitting author:** bhazelton (<a href="http://orcid.org/0000-0001-7532-645X">Bryna J. Hazelton</a>)
**Repository:** https://github.com/RadioAstronomySoftwareGroup/pyradiosky
**Branch with paper.md** (empty if default branch):
**Version:** 1.0
