### 2024-03-18 05ef4a1
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/psf/black: 24.2.0 → 24.3.0](https://github.com/psf/black/compare/24.2.0...24.3.0)
