### 2024-03-26 RadioAstronomySoftwareGroup/pyradiosky#238 @pritchardn
Inconsistent License Description
Hello, this is a minor issue relating to the [submission in JOSS](https://github.com/openjournals/joss-reviews/issues/6503)

The [license file](https://github.com/RadioAstronomySoftwareGroup/pyradiosky/blob/main/LICENSE) at the top level is BSD 2-clause, but several of the code files state it is licensed under the 3-clause variant (e.g. [skymodel.py](https://github.com/RadioAstronomySoftwareGroup/pyradiosky/blob/main/src/pyradiosky/skymodel.py)).
While perhaps a bit nit-picky, this is easy to fix. :smile:
