### 2024-03-26 #6503 @pritchardn
While the license file is valid, the code-files refer to the BSD-3 clause variant - this is a pretty easy to fix, but probably worth getting consistent. I've made [an issue](https://github.com/RadioAstronomySoftwareGroup/pyradiosky/issues/238) to address this.
