+++ setup.py
@@ -26,9 +26,9 @@ def branch_scheme(version):
 with io.open("README.md", "r", encoding="utf-8") as readme_file:
     readme = readme_file.read()
 
-healpix_reqs = ["astropy-healpix>=0.6"]
+healpix_reqs = ["astropy-healpix>=1.0.2"]
 astroquery_reqs = ["astroquery>=0.4.4"]
-lunarsky_reqs = ["lunarsky>=0.2.1"]
+lunarsky_reqs = ["lunarsky>=0.2.2"]
 all_optional_reqs = healpix_reqs + astroquery_reqs + lunarsky_reqs
 
 doc_reqs = ["matplotlib", "pypandoc", "sphinx"]
@@ -52,11 +52,12 @@ setup_args = {
     "use_scm_version": {"local_scheme": branch_scheme},
     "include_package_data": True,
     "install_requires": [
-        "numpy>=1.20",
-        "scipy>=1.5",
-        "astropy>=5.2",
-        "h5py>=3.1",
-        "pyuvdata>=2.4.1",
+        "astropy>=6.0",
+        "h5py>=3.4",
+        "numpy>=1.23",
+        "pyuvdata>=2.4.3",
+        "scipy>=1.7.3",
+        "setuptools>=61",
         "setuptools_scm>=7.0.3",
     ],
     "extras_require": {
@@ -72,7 +73,6 @@ setup_args = {
         "Development Status :: 1 - Planning",
         "Intended Audience :: Science/Research",
         "License :: OSI Approved :: BSD License",
-        "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
         "Programming Language :: Python :: 3.11",
         "Programming Language :: Python :: 3.12",
