+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Changed
+- Updated minimum dependency versions: astropy>=6.0, h5py>=3.4, numpy>=1.23,
+pyuvdata>=2.4.3, scipy>=1.7.3, python>=3.10
+- Updated minimum optional dependency versions: astropy-healpix>=1.0.2, lunarsky>=0.2.2
+
 ## [0.3.1] - 2024-02-22
 
 ### Added
