+++ ci/full_deps.yaml
@@ -2,17 +2,17 @@ name: full_deps
 channels:
   - conda-forge
 dependencies:
-  - astropy>=5.2
-  - astropy-healpix>=0.6
+  - astropy>=6.0
+  - astropy-healpix>=1.0.2
   - astroquery>=0.4.4
   - coverage
-  - h5py>=3.0
+  - h5py>=3.4
   - matplotlib  # just for the doctests
-  - numpy>=1.20
+  - numpy>=1.23
   - pip
   - pytest-cov
-  - pyuvdata>=2.4.1
-  - scipy>=1.5
+  - pyuvdata>=2.4.3
+  - scipy>=1.7.3
   - setuptools_scm>=7.0.3
   - pip:
-      - lunarsky>=0.2.1
+      - lunarsky>=0.2.2
