+++ ci/min_deps.yaml
@@ -2,13 +2,13 @@ name: min_deps
 channels:
   - conda-forge
 dependencies:
-  - astropy>=5.2
+  - astropy>=6.0
   - coverage
-  - h5py>=3.0
-  - numpy>=1.20
+  - h5py>=3.4
+  - numpy>=1.23
   - pytest
   - pytest-cov
-  - pyuvdata>=2.4.1
-  - scipy>=1.5
+  - pyuvdata>=2.4.3
+  - scipy>=1.7.3
   - setuptools_scm>=7.0.3
   - pip
