+++ .github/workflows/testsuite.yaml
@@ -21,7 +21,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: [3.9, "3.10", "3.11", "3.12"]
+        python-version: ["3.10", "3.11", "3.12"]
     steps:
       - uses: actions/checkout@main
         with:
@@ -123,7 +123,7 @@ jobs:
   min_versions:
     env:
       ENV_NAME: min_versions
-      PYTHON: 3.9
+      PYTHON: "3.10"
     name: Min Versions Testing
     runs-on: ubuntu-latest
     strategy:
@@ -178,7 +178,7 @@ jobs:
   # Use pip for diversity
   warning_test:
     env:
-      PYTHON: "3.10"
+      PYTHON: "3.11"
     name: Warning Test
     runs-on: ubuntu-latest
     steps:
