+++ ci/min_versions.yaml
@@ -2,17 +2,17 @@ name: min_versions
 channels:
   - conda-forge
 dependencies:
-  - astropy==5.2
-  - astropy-healpix==0.6
+  - astropy==6.0
+  - astropy-healpix==1.0.2
   - astroquery==0.4.4
-  - h5py==3.1.*
+  - h5py==3.4.0
   # should be 1.20.* but the old packages of pyuvdata accidentally required 1.23
-  - numpy==1.23.*
-  - scipy==1.5.*
+  - numpy==1.23
+  - scipy==1.7.3
   - coverage
   - pytest-cov
   - setuptools_scm==7.0.3
-  - pyuvdata==2.4.1
+  - pyuvdata==2.4.3
   - pip
   - pip:
-      - lunarsky==0.2.1
+      - lunarsky==0.2.2
