+++ ci/publish.yaml
@@ -2,12 +2,12 @@ name: publish
 channels:
   - conda-forge
 dependencies:
-  - astropy>=5.2
-  - h5py>=3.0
-  - numpy>=1.20
+  - astropy>=6.0
+  - h5py>=3.4
+  - numpy>=1.23
   - pip
-  - pyuvdata>=2.4.1
-  - scipy>=1.5
+  - pyuvdata>=2.4.3
+  - scipy>=1.7.3
   - setuptools_scm>=7.0.3
   - pip:
     - build
