### 2024-03-28 bcb0431
handle occasional vizier download errors
