### 2024-03-28 61a1e98
fix minor doc bug: update_positions can take a MoonLocation too
