+++ .pre-commit-config.yaml
@@ -2,7 +2,7 @@ exclude: '(^docs/conf.py|^user_data/External_tables/)'
 
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.5.0
+    rev: v4.6.0
     hooks:
         - id: trailing-whitespace
         - id: check-added-large-files
@@ -29,7 +29,7 @@ repos:
           - pep8-naming
 
   - repo: https://github.com/psf/black
-    rev: 24.3.0
+    rev: 24.4.0
     hooks:
         - id: black
 
