+++ .github/workflows/externaltests.yaml
@@ -54,4 +54,4 @@ jobs:
           cd pyuvsim
           pip install --no-deps .
           mkdir test-reports
-          python -m pytest pyuvsim --junitxml=test-reports/xunit.xml
+          python -m pytest --junitxml=test-reports/xunit.xml
