### 2024-04-26 #6503 @bhazelton
> `pytest` runs with the following:
>
> ```
> tests/test_skymodel.py: 618 warnings
>   `product` is deprecated as of NumPy 1.25.0, and will be removed in NumPy 2.0. Please use `prod` instead.
> ```
>
> @bhazelton -- Minor thing if you wish to resolve; otherwise the tests runs fine.

@0xCoto I can’t seem to replicate these warnings. I made a fresh environment with all the dependencies and I don’t get them. I can’t find anywhere in pyradiosky or pyuvdata (a dependency that I help maintain) that calls np.prod, so I think these warnings must be coming from another dependency and I must have a different version of that dependency than you do.
