+++ .pre-commit-config.yaml
@@ -29,7 +29,7 @@ repos:
           - pep8-naming
 
   - repo: https://github.com/psf/black
-    rev: 24.4.0
+    rev: 24.4.2
     hooks:
         - id: black
 
