### 2024-04-30 71ef2dd
handle transient spiceypy errors
