+++ paper.bib
@@ -49,3 +49,46 @@
   journal = {GitHub repository},
   url = {https://github.com/aelanman/lunarsky}
 }
+
+
+@inproceedings{topcat,
+    author = {{Taylor}, M.~B.},
+    title = "{TOPCAT \& STIL: Starlink Table/VOTable Processing Software}",
+    booktitle = {Astronomical Data Analysis Software and Systems XIV},
+    year = 2005,
+    editor = {{Shopbell}, P. and {Britton}, M. and {Ebert}, R.},
+    series = {Astronomical Society of the Pacific Conference Series},
+    volume = {347},
+    month = dec,
+    pages = {29},
+    adsurl = {https://ui.adsabs.harvard.edu/abs/2005ASPC..347...29T},
+    adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+@ARTICLE{healpix,
+       author = {{G{\'o}rski}, K.~M. and {Hivon}, E. and {Banday}, A.~J. and {Wandelt}, B.~D. and {Hansen}, F.~K. and {Reinecke}, M. and {Bartelmann}, M.},
+        title = "{HEALPix: A Framework for High-Resolution Discretization and Fast Analysis of Data Distributed on the Sphere}",
+      journal = {The Astrophysical Journal},
+     keywords = {Cosmology: Cosmic Microwave Background, Cosmology: Observations, Methods: Statistical, Astrophysics},
+         year = 2005,
+        month = apr,
+       volume = {622},
+       number = {2},
+        pages = {759-771},
+          doi = {10.1086/427976},
+archivePrefix = {arXiv},
+       eprint = {astro-ph/0409513},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2005ApJ...622..759G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@misc{astropy-healpix,
+  author = {{Astropy Developers}},
+  title = {astropy healpix: BSD-licensed HEALPix for Astropy},
+  year = {2016},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/astropy/astropy-healpix}
+}
