### 2024-05-02 #6503 @bhazelton
> Everything looks OK to me. The only thing I would request is a brief mention of how the software compares to other commonly-used packages, so that readers understand whether their needs are covered by `pyradiosky` vs a different module.

@0xCoto We have updated our paper with a paragraph addressing this.
