### 2024-05-06 #6503 @xuanxu
@bhazelton looks like we're close to being done here. Please give your own paper a final read to check for any potential typos etc.
A small comment on one of your references: You cite `astropy` using their first paper from 2013, but the Astropy's guidelines [recommend](https://www.astropy.org/acknowledging.html#acknowledging-or-citing-astropy) citing other papers or at least the more recent one, please consider updating the paper with that suggestion.
