### 2024-05-06 1d48140
Fix the astropy references per editor's request
