### 2024-05-06 a6f3cb3
Fix the astropy references per editor's request
