+++ paper.bib
@@ -26,20 +26,164 @@
     journal = {Journal of Open Source Software}
 }
 
-@article{astropy,
-    author = {{Astropy Collaboration}},
-    title = "{Astropy: A community Python package for astronomy}",
-    journal = {Astronomy and Astrophysics},
-    archivePrefix = "arXiv",
-    eprint = {1307.6212},
-    primaryClass = "astro-ph.IM",
-    keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
-    year = 2013,
-    month = oct,
-    volume = 558,
-    doi = {10.1051/0004-6361/201322068},
-    url = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A}
+
+@article{astropy:2013,
+    Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+    Adsurl = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A},
+    Archiveprefix = {arXiv},
+    Author = {{Astropy Collaboration} and {Robitaille}, T.~P. and {Tollerud}, E.~J.
+        and {Greenfield}, P. and {Droettboom}, M. and {Bray}, E. and {Aldcroft}, T.
+        and {Davis}, M. and {Ginsburg}, A. and {Price-Whelan}, A.~M. and {Kerzendorf}, W.~E.
+        and {Conley}, A. and {Crighton}, N. and {Barbary}, K. and {Muna}, D.
+        and {Ferguson}, H. and {Grollier}, F. and {Parikh}, M.~M. and {Nair}, P.~H.
+        and {Unther}, H.~M. and {Deil}, C. and {Woillez}, J. and {Conseil}, S.
+        and {Kramer}, R. and {Turner}, J.~E.~H. and {Singer}, L. and {Fox}, R.
+        and {Weaver}, B.~A. and {Zabalza}, V. and {Edwards}, Z.~I.
+        and {Azalee Bostroem}, K. and {Burke}, D.~J. and {Casey}, A.~R.
+        and {Crawford}, S.~M. and {Dencheva}, N. and {Ely}, J. and {Jenness}, T.
+        and {Labrie}, K. and {Lim}, P.~L. and {Pierfederici}, F. and {Pontzen}, A.
+        and {Ptak}, A. and {Refsdal}, B. and {Servillat}, M. and {Streicher}, O.
+    },
+    Doi = {10.1051/0004-6361/201322068},
+    Eid = {A33},
+    Eprint = {1307.6212},
+    Journal = {\aap},
+    Keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
+    Month = oct,
+    Pages = {A33},
+    Primaryclass = {astro-ph.IM},
+    Title = {{Astropy: A community Python package for astronomy}},
+    Volume = 558,
+    Year = 2013,
+    Bdsk-Url-1 = {https://dx.doi.org/10.1051/0004-6361/201322068}
 }
+
+@article{astropy:2018,
+    author = {{Astropy Collaboration} and {Price-Whelan}, A.~M. and
+        {Sip{\H{o}}cz}, B.~M. and {G{\"u}nther}, H.~M. and {Lim}, P.~L. and
+        {Crawford}, S.~M. and {Conseil}, S. and {Shupe}, D.~L. and
+        {Craig}, M.~W. and {Dencheva}, N. and {Ginsburg}, A. and {VanderPlas}, J.~T.
+        and {Bradley}, L.~D. and {P{\'e}rez-Su{\'a}rez}, D. and
+        {de Val-Borro}, M. and {Aldcroft}, T.~L. and {Cruz}, K.~L. and
[…]
