### 2024-05-09 f609e03
Minor paper wording updates.
