### 2024-05-09 6947e25
Fix a couple things that cause warnings in astropy 6.1
