### 2024-05-14 #6503 @bhazelton
Yes, sorry. The version is 1.0 (https://github.com/RadioAstronomySoftwareGroup/pyradiosky/releases/tag/v1.0.0) and the zenodo DOI is 10.5281/zenodo.11187469 (https://doi.org/10.5281/zenodo.11187469)
