+++ paper.bib
@@ -47,7 +47,7 @@
     Doi = {10.1051/0004-6361/201322068},
     Eid = {A33},
     Eprint = {1307.6212},
-    Journal = {\aap},
+    Journal = {Astronomy and Astrophysics},
     Keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
     Month = oct,
     Pages = {A33},
@@ -104,7 +104,7 @@
         {Zabalza}, V. and {Astropy Contributors}
     },
     title = "{The Astropy Project: Building an Open-science Project and Status of the v2.0 Core Package}",
-    journal = {\aj},
+    journal = {Astronomical Journal},
     keywords = {methods: data analysis, methods: miscellaneous, methods: statistical, reference systems, Astrophysics - Instrumentation and Methods for Astrophysics},
     year = 2018,
     month = sep,
@@ -168,7 +168,7 @@
         and {Zonca}, Andrea and {Astropy Project Contributors}
     },
     title = "{The Astropy Project: Sustaining and Growing a Community-oriented Open-source Project and the Latest Major Release (v5.0) of the Core Package}",
-    journal = {\apj},
+    journal = {Astrophysical Journal},
     keywords = {Astronomy software, Open source software, Astronomy data analysis, 1855, 1866, 1858, Astrophysics - Instrumentation and Methods for Astrophysics},
     year = 2022,
     month = aug,
@@ -186,7 +186,7 @@
 
 
 @misc{lunarsky,
-  author = {Adam Lanman},
+  author = {Adam E. Lanman},
   title = {lunarsky: An astropy extension to describe observations from the surface of the Moon.},
   year = {2020},
   publisher = {GitHub},
